First release of PPINot4Py, big structure change, configuration for pypi packaging. Changes in imports and package system
