+++ PPINot4Py/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\ErMamaweboPlus\\Miniconda3\\python.exe"
+}
