+++ PPINot4Py/computers/countComputer.py
@@ -1,0 +1,11 @@
+from PPINot4Py.condition.conditionChooser import conditionChooser
+
+def countCompute(dataframe, condition, id_case):
+
+    precondition = (condition.when)
+
+    filtered_series = conditionChooser(dataframe, id_case, precondition)
+
+    result = filtered_series.groupby(dataframe[id_case]).sum()
+
+    return result
