+++ PPINot4Py/condition/Condition.py
