+++ PPINot4Py/condition/DataPropertyCondition.py
@@ -1,0 +1,7 @@
+import PPINot4Py.condition.ProcessInstanceCondition as ProcessInstanceCondition
+
+#Possibly not needed
+class DataPropertyCondition(ProcessInstanceCondition):
+    
+      def __init__(self):
+         super(DataPropertyCondition, self).__init__()
