+++ PPINot4Py/condition/ProcessInstanceCondition.py
@@ -1,0 +1,4 @@
+class ProcessInstanceCondition():
+
+    def __init__(self):
+         super(ProcessInstanceCondition, self).__init__()
