+++ PPINot4Py/functionsFrame/__init__.py
@@ -1,0 +1,1 @@
+from PPINot4Py.functionsFrame import auxFunctions
