+++ PPINot4Py/latexDocuments/example.aux
@@ -1,0 +1,9 @@
+\relax 
+\@writefile{toc}{\contentsline {section}{\numberline {1}Measures}{1}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsection}{\numberline {1.1}CountMeasure}{1}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsection}{\numberline {1.2}DataMeasure}{1}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsection}{\numberline {1.3}TimeMeasure}{1}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsubsection}{\numberline {1.3.1}Linear type}{2}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsubsection}{\numberline {1.3.2}Cyclic type}{2}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsubsection}{\numberline {1.3.3}AggFuntions}{2}\protected@file@percent }
+\@writefile{toc}{\contentsline {subsection}{\numberline {1.4}aggregatedMeasure}{2}\protected@file@percent }
