+++ PPINot4Py/latexDocuments/example.tex
@@ -1,0 +1,148 @@
+% !TEX TS-program = pdflatex
+% !TEX encoding = UTF-8 Unicode
+
+% This is a simple template for a LaTeX document using the "article" class.
+% See "book", "report", "letter" for other types of document.
+
+\documentclass[11pt]{article} % use larger type; default would be 10pt
+
+\usepackage[utf8]{inputenc} % set input encoding (not needed with XeLaTeX)
+
+%%% Examples of Article customizations
+% These packages are optional, depending whether you want the features they provide.
+% See the LaTeX Companion or other references for full information.
+
+%%% PAGE DIMENSIONS
+\usepackage{geometry} % to change the page dimensions
+\geometry{a4paper} % or letterpaper (US) or a5paper or....
+% \geometry{margin=2in} % for example, change the margins to 2 inches all round
+% \geometry{landscape} % set up the page for landscape
+%   read geometry.pdf for detailed page layout information
+
+\usepackage{graphicx} % support the \includegraphics command and options
+
+% \usepackage[parfill]{parskip} % Activate to begin paragraphs with an empty line rather than an indent
+
+%%% PACKAGES
+\usepackage{booktabs} % for much better looking tables
+\usepackage{array} % for better arrays (eg matrices) in maths
+\usepackage{paralist} % very flexible & customisable lists (eg. enumerate/itemize, etc.)
+\usepackage{verbatim} % adds environment for commenting out blocks of text & for better verbatim
+\usepackage{subfig} % make it possible to include more than one captioned figure/table in a single float
+% These packages are all incorporated in the memoir class to one degree or another...
+
+\usepackage{listings}
+\usepackage{xcolor}
+ 
+\definecolor{codegreen}{rgb}{0,0.6,0}
+\definecolor{codegray}{rgb}{0.5,0.5,0.5}
+\definecolor{codepurple}{rgb}{0.58,0,0.82}
+\definecolor{backcolour}{rgb}{0.95,0.95,0.92}
+
+\lstset{showstringspaces=false}
+\lstdefinestyle{Python}{
+    language        = Python,
+    basicstyle      = \ttfamily,
+    numbers=left,
+    stepnumber=1,
+    keywordstyle    = \color{blue},
+    stringstyle     = \color{red},
+    commentstyle    = \color{red}\ttfamily
+}
+
+%%% HEADERS & FOOTERS
+\usepackage{fancyhdr} % This should be set AFTER setting up the page geometry
+\pagestyle{fancy} % options: empty , plain , fancy
+\renewcommand{\headrulewidth}{0pt} % customise the layout...
+\lhead{}\chead{}\rhead{}
+\lfoot{}\cfoot{\thepage}\rfoot{}
+
+%%% SECTION TITLE APPEARANCE
+\usepackage{sectsty}
+\allsectionsfont{\sffamily\mdseries\upshape} % (See the fntguide.pdf for font help)
+% (This matches ConTeXt defaults)
+
+%%% ToC (table of contents) APPEARANCE
+\usepackage[nottoc,notlof,notlot]{tocbibind} % Put the bibliography in the ToC
+\usepackage[titles,subfigure]{tocloft} % Alter the style of the Table of Contents
+\renewcommand{\cftsecfont}{\rmfamily\mdseries\upshape}
+\renewcommand{\cftsecpagefont}{\rmfamily\mdseries\upshape} % No bold!
+
+%%% END Article customizations
+
+%%% The "real" document content comes below...
+
+\title{PPINot4Py description}
+\author{Alejandro Gómez Caballero}
+%\date{} % Activate to display a given date or no date (if empty),
+         % otherwise the current date is printed 
+
+\begin{document}
+\maketitle
+
+\section{Measures}
+
+Defining a measure is the way to indicate wich kind of operation we want to do in our dataframe, we have several tipes:
+
+
+\subsection{CountMeasure}
+Most basic measure, used to count the times a condition occurs.\\
+It can receive as parameter a String condition o a TimeInstantCondition
+
+\begin{lstlisting}[style = Python]
+countStateTimeA = 
+    DataObjectState('lifecycle_transition == "In Progress"')
+countConditionTimeA = 
+    TimeInstantCondition(countStateTimeA)
+countMeasureTimeA = 
+    CountMeasure(countConditionTimeA)
+
+#----------------------------'OR'---------------------------
+
+countMeasureTimeA = 
+    CountMeasure('lifecycle_transition == "In Progress"')    
+
+\end{lstlisting}
+
+\subsection{DataMeasure}
+With this measure we can obtain the first or last appareance of a value when we apply a TimeInstantCondittion
+
+\subsection{TimeMeasure}
+
+In this measure, we want to count how much time has happened between 1 event and other, in a linear form or in a cyclic form\\
+
+If we call "A" to the first condition and "B" to the second, we want the appareances of the pairs A$\,\to\,$B\\
+We have the following parameters:\\
+\indent
+FromCondition: The first TimeInstantCondition \\
+\indent
+ToCondition: The second TimeInstantCondition \\
+\indent
+TimeMeasureType: Tipe of calculation, Linear or Cyclic.\\
+\indent
+SingleInstanceAggFunction: Tipe of operation we want to apply.\\
+\indent
+Precondition: In case we want to apply a previous filter to our dataframe\\ 
+\indent
+FirstTo: Only works for the linear type, is True if you want the first "B" and False if you want the last "B"\\
+\subsubsection{Linear type}
+In this case, we are counting the first appareance of A condition and the first/last appareance of B condition, then we calculate the time difference between B and A\\
+\subsubsection{Cyclic type}
+In cyclic case, we count each apparition of A$\,\to\,$B apparitions and calculate an operation to this apparitions.
+\subsubsection{AggFuntions}
+We have 5 possible operations by defect in the program:\\
+
+SUM: The sum of the time values of all A$\,\to\,$B apparitions\\
+\indent
+MAX: Max time value between the A$\,\to\,$B pairs\\
+\indent
+MIN: Min time value between the A$\,\to\,$B pairs\\
+\indent
+AVG: The average of the time values of all A$\,\to\,$B apparitions\\
+\indent
+GROUPBY: Raw grouped dataframe with no operation applied\\
+
+\subsection{aggregatedMeasure}
+
+
+\end{document}
