+++ PPINot4Py/state/ComplexState.py
@@ -1,0 +1,18 @@
+class ComplexState():
+    
+        def __init__(self):
+            self.first = TimeInstantCondition.TimeInstantCondition()
+            self.last = TimeInstantCondition.TimeInstantCondition()
+            self.type = Type
+        
+
+        def getFirst(self):
+            return self.first
+
+
+        def getLast(self):
+            return self.last
+        
+
+        def getType(self):
+            return self.type
