--- model/condition/ConditionImpl.py
@@ -1,8 +1,0 @@
-#Possibly not needed
-class ConditionImpl():
-
-    def __init__(self, appliesTo):
-        self.appliesTo = appliesTo
-
-    def getAppliesTo(self):
-        return self.appliesTo
