--- model/condition/DataPropertyCondition.py
@@ -1,7 +1,0 @@
-import condition.ProcessInstanceCondition as ProcessInstanceCondition
-
-#Possibly not needed
-class DataPropertyCondition(ProcessInstanceCondition):
-    
-      def __init__(self):
-         super(DataPropertyCondition, self).__init__()
