--- model/condition/ProcessInstanceCondition.py
@@ -1,6 +1,0 @@
-import condition.ConditionImpl as ConditionImpl
-#Possibly not needed
-class ProcessInstanceCondition(ConditionImpl):
-
-    def __init__(self):
-         super(ProcessInstanceCondition, self).__init__()
