--- model/jsonTest.json
@@ -1,11 +1,0 @@
-{
-    "when": {
-        "changesToState": {
-            "dataObjectState": {
-                "name": "lifecycle_transition == 'In Progress'"
-                                }
-                            }
-            }
-}
-
-
