--- model/state/ComplexState.py
@@ -1,20 +1,0 @@
-import state as RunTimeState 
-
-class ComplexState(RuntimeState):
-    
-        def __init__(self):
-            self.first = TimeInstantCondition.TimeInstantCondition()
-            self.last = TimeInstantCondition.TimeInstantCondition()
-            self.type = Type
-        
-
-        def getFirst(self):
-            return self.first
-
-
-        def getLast(self):
-            return self.last
-        
-
-        def getType(self):
-            return self.type
