--- model/state/DataObjectState.py
@@ -1,16 +1,0 @@
-import state as RunTimeState
-
-class DataObjectState(RunTimeState):
-    
-    def __init__(self, name):
-        #super(DataObjectState, self).__init__()
-        self.name = name
-        
-    def getName(self):
-        return self.__name
-        
-    def setName(self):
-        self.name = name
-
-    def __repr__(self):
-        return "%s" % (self.name)
